project/backend/requirements.txt
fastapi==0.115.8
uvicorn==0.19.0
httpx==0.28.0
python-dotenv==1.0.1
openai==0.27.8
numpy==2.2.2
pandas==2.2.3
pymysql==1.1.1
sqlalchemy==2.0.20